Update to rename to createGlobalContext API
