+++ .eslintrc.js
@@ -12,8 +12,6 @@ module.exports = {
     'plugin:@typescript-eslint/recommended',
     'plugin:import/typescript',
     'prettier',
-    'prettier/@typescript-eslint',
-    'prettier/react',
   ],
   globals: {
     Atomics: 'readonly',
