+++ .lint-staged.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  '*': [() => 'yarn format', () => 'yarn test'],
+};
