--- .lintstagedrc
@@ -1,6 +1,0 @@
-{
-  "*.{ts,tsx,js,jsx,json}": [
-    "prettier --write",
-    "jest --bail --findRelatedTests"
-  ]
-}
