+++ setupTests.js
@@ -1,5 +1,6 @@
-// setup file
-import { configure } from 'enzyme';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+/* eslint-disable @typescript-eslint/no-var-requires */
+
+const { configure } = require('enzyme');
+const Adapter = require('@wojtekmaj/enzyme-adapter-react-17');
 
 configure({ adapter: new Adapter() });
