+++ src/createGlobalContext/index.ts
@@ -1,10 +1,6 @@
-// createGlobalContext API
 export { createGlobalContext } from './createGlobalContext';
 export type {
   CreateGlobalContextArgs,
   GlobalContextProviderProps,
   UseReducerContextValue,
 } from './createGlobalContext';
-
-// createStore API
-export { createStore } from './createStore';
