--- src/createUseReducerContext/index.ts
@@ -1,2 +1,0 @@
-export { createUseReducerContext } from './createUseReducerContext';
-export type { UseReducerContextValue } from './createContext';
