--- src/createUseStateContext/index.ts
@@ -1,2 +1,0 @@
-export { createUseStateContext } from './createUseStateContext';
-export type { UseStateContextValue } from './createContext';
