+++ src/types.ts
@@ -1,3 +1,6 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/* @ts-ignore */
 import react from 'react';
 
 declare module 'react' {
