+++ tsconfig.json
@@ -5,7 +5,7 @@
     "target": "ESNext",
     "module": "ESNext",
     "lib": ["ESNext", "dom"],
-    "jsx": "preserve",
+    "jsx": "react-jsx",
     "allowJs": false,
     "checkJs": false,
     "skipLibCheck": true,
